backend/pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "LucasGabrielAntonete", email = "lucasantonete@gmail.com"},
]
dependencies = [
    "django>=4.2.7",
    "setuptools>=69.0.2",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.3.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

